Workaround -Wno-* error in with Google GCC 4.4.3
Seems that some versions of GCC don't report missing -Wno-*
flags correctly. e.g. gcc -Wall -Werror -Wno-way-this-should-work
 will not report an error nor will it return a non-zero exit code

If the flag didn't have no- as the first letters then we'd get
the result we expect. For this reason, check for the positive
version of each warning to be disabled and then string replace
it to get the negative (-Wno-*) version afterwards.

Some other minor tidyups (removing some HAVE_PRIVEXT nonsense)