Removed most of the AH_VERBATIM nonsense.
I say nonsense as it was declarations directly in configure.ac.
They are now in emulated.h where they should be. There is still
one line of AH_VERBATIM to include emulated.h into config.h.

config.h is then -include'd on the command line to minimize
changes to the original source code.